requirements.txt
pandas
numpy
scikit-learn==1.3.2
streamlit==1.31.1
Pillow
joblib
